Moved release notes over to fragment, added motion layout title
