+++ app/src/main/res/layout/activity_release_notes.xml
@@ -1,90 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/f1BackgroundPrimary">
 
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@null"
-        app:elevation="0dp"
-        app:layout_behavior="com.google.android.material.appbar.AppBarLayout$Behavior">
-
-        <com.google.android.material.appbar.CollapsingToolbarLayout
-            android:id="@+id/collapsingtoolbar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_scrollFlags="exitUntilCollapsed|scroll|enterAlways">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:id="@+id/titlebar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:background="?attr/f1BackgroundPrimary"
-                android:fitsSystemWindows="false">
-
-                <TextView
-                    android:id="@+id/header"
-                    style="@style/Text.Toolbar"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:fitsSystemWindows="true"
-                    android:gravity="center"
-                    android:paddingTop="@dimen/paddingMedium"
-                    android:paddingBottom="@dimen/paddingMedium"
-                    android:text="@string/settings_help_release_notes_title"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <ImageButton
-                    android:id="@+id/back"
-                    android:layout_width="@dimen/toolbarSettingsIcon"
-                    android:layout_height="@dimen/toolbarSettingsIcon"
-                    android:layout_marginStart="@dimen/paddingSmall"
-                    android:layout_marginTop="@dimen/paddingSmall"
-                    android:layout_marginEnd="@dimen/paddingMedium"
-                    android:layout_marginBottom="@dimen/paddingSmall"
-                    android:background="@null"
-                    android:contentDescription="@string/settings_title"
-                    android:scaleType="center"
-                    android:src="@drawable/ic_back"
-                    android:tint="?attr/f1TextPrimary"
-                    app:layout_constraintBottom_toBottomOf="@id/header"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="@id/header" />
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-        </com.google.android.material.appbar.CollapsingToolbarLayout>
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <androidx.core.widget.NestedScrollView
-        android:id="@+id/scrollview"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragments"
+        android:name="tmg.flashback.ui.settings.release.ReleaseFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:clipToPadding="false"
-        android:overScrollMode="never"
-        app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView xmlns:tools="http://schemas.android.com/tools"
-            android:id="@+id/tvReleaseNotes"
-            style="@style/Text.ReleaseNotes.Content"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingStart="@dimen/paddingMedium"
-            android:paddingTop="@dimen/paddingMedium"
-            android:paddingEnd="@dimen/paddingMedium"
-            android:paddingBottom="@dimen/paddingMedium"
-            tools:text="@tools:sample/lorem" />
-
-    </androidx.core.widget.NestedScrollView>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
