+++ app/src/main/res/layout/fragment_release_notes.xml
@@ -3,70 +3,98 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="?attr/f1BackgroundPrimary"
+    android:layout_height="match_parent"
     android:orientation="vertical">
+    <!-- Linear Layout parent required because of DiscordOverlappingPanels Weirdness -->
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <View
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="0dp" />
 
-        <View
-            android:id="@+id/pill"
-            android:layout_width="@dimen/bottomSheetPillWidth"
-            android:layout_height="@dimen/bottomSheetPillHeight"
-            android:layout_gravity="center"
-            android:layout_marginTop="@dimen/paddingSmall"
-            android:background="?attr/f1BackgroundSecondary"
-            app:layout_constraintEnd_toEndOf="parent"
+    <androidx.constraintlayout.motion.widget.MotionLayout
+        android:id="@+id/container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="?attr/rssBackground"
+        app:layoutDescription="@xml/scene_fragment_release_notes"
+        tools:showPaths="true">
+
+        <androidx.constraintlayout.widget.Guideline
+            android:id="@+id/toolbarGuideline"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            app:layout_constraintGuide_begin="@dimen/toolbarHeight" />
+
+        <ImageView
+            android:id="@+id/back"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/paddingSmall"
+            android:clickable="true"
+            android:contentDescription="@string/about_this_app_ab_back_button"
+            android:focusable="true"
+            android:foreground="@drawable/rounded_ripple"
+            android:padding="@dimen/paddingSmall"
+            app:layout_constraintBottom_toBottomOf="@id/toolbarGuideline"
             app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@drawable/ic_back"
+            app:tint="?attr/f1TextPrimary" />
+
+        <TextView
+            android:id="@+id/titleCollapsed"
+            style="@style/Text.Toolbar.Collapsed"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/paddingSmall"
+            android:layout_marginEnd="@dimen/paddingMedium"
+            android:alpha="0.0"
+            android:text="@string/settings_help_release_notes_title"
+            app:layout_constraintBottom_toBottomOf="@id/toolbarGuideline"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="@id/back"
             app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
-            android:id="@+id/bottomSheetTitle"
-            style="@style/Text.BottomSheet.Title"
+            android:id="@+id/titleExpanded"
+            style="@style/Text.Toolbar.Expanded"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="@dimen/paddingSmall"
-            android:gravity="center"
-            android:text="@string/settings_help_release_notes_title"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/pill" />
-
-        <TextView
-            android:id="@+id/subtitle"
-            style="@style/Text.BottomSheet.Subtitle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
+            android:layout_marginTop="@dimen/toolbarExpandedSpacer"
             android:paddingStart="@dimen/paddingMedium"
             android:paddingTop="@dimen/paddingSmall"
             android:paddingEnd="@dimen/paddingMedium"
-            android:paddingBottom="@dimen/paddingNSmall"
-            android:text="@string/release_notes_description"
-            android:gravity="center"
+            android:paddingBottom="@dimen/paddingMedium"
+            android:text="@string/settings_help_release_notes_title"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/bottomSheetTitle" />
+            app:layout_constraintTop_toTopOf="@id/toolbarGuideline" />
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+        <androidx.core.widget.NestedScrollView
+            android:id="@+id/content"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:overScrollMode="never"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/titleExpanded">
 
-    <View
-        android:layout_width="wrap_content"
-        android:layout_height="1dp"
-        android:background="?attr/f1BackgroundSecondary" />
+            <TextView
+                android:id="@+id/releaseNotes"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:clipToPadding="false"
+                android:overScrollMode="never"
+                android:paddingStart="@dimen/paddingMedium"
+                android:paddingTop="@dimen/paddingSmall"
+                android:paddingEnd="@dimen/paddingMedium"
+                android:paddingBottom="@dimen/paddingXLarge"
+                android:textColor="?attr/f1TextPrimary" />
 
+        </androidx.core.widget.NestedScrollView>
 
-    <TextView
-        android:id="@+id/tvReleaseNotesDescription"
-        style="@style/Text.ReleaseNotes.Content"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/paddingMedium"
-        android:paddingTop="@dimen/paddingSmall"
-        android:paddingEnd="@dimen/paddingMedium"
-        android:paddingBottom="@dimen/paddingXLarge"
-        tools:text="Initial release!" />
+    </androidx.constraintlayout.motion.widget.MotionLayout>
 
 </LinearLayout>
